Laguerre root solver
This commit includes the code of a root solver using the Laguerre's
method. The method was embedded on a class so as to be used in other cpp
projects.

It contains a core function where the Laguerre method is coded and a
driver routine that removes from the polynomial the roots as they are
found.